+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 ## 2016-05-24, Version 0.4.0
 
 You can send request as a json or text
+
+## 201-01-20, Version 0.7.0
+
+Rename `HttpRequest` to `CurlRequest`
+
+Rename `HttpResponse` to `CurlResponse`